Add script for copying terminfo while SSHing
This is a bit of a hack. The script only tries to copy the terminfo with
simple ssh commands ('ssh HOST'), and when it does this SSH thinks the
session is not interactive, so the script has to do stuff like print out
the MOTD itself. It does seem to generally work, though.

I think I will try another approach soon in which I use `expect` to make
a normal interactive connection and then send the tic command after the
prompt appears.