[tmux] Bind " and %; use correct flags when setting options
As far as I can tell, it doesn't actually matter whether you use the
correct flags (-s for server options / -w for window options) when
setting the options, but you do have to pass the right flags to
show-options to *see* those options.  So I'm bothering to use the right
flags here to make it easier for myself to debug things later.