[bashrc] Print newline (and indicator) before prompt when needed
This queries the terminal at the start of PS1 evaluation to find the
current column of the cursor. If the cursor is not in the first column,
it prints an indicator and a newline. The result is that the prompt
always starts on a new line, but I can still tell if the output didn't
end with a newline thanks to the indicator.

Although, if a file ends with no newline but the cursor happens to wrap
exactly back to the first column, this hack won't know the difference.