Add userContent.css as a template
In my experience, Firefox won't load this file if it is a symlink.
Making it a template seemed easier than supporting both symlinking and
copying in the non-template flow (which already has mtime safety
checks). Also, it seems likely that this file will need to be different
for different machines in the future.