Add script and configs for a global push-to-talk key
The key is F13, which apparently is known as XF86Tools here. I have a
button on my mouse as well as a key on my keyboard mapped to F13. I
accomplish the push-to-talk by muting and unmuting the default
PulseAudio source, and I display the status in the top center of my
screen.