Enable focus events in tmux; fix pasting
Now that I have some tmux bindings that sometimes appear in pasted text,
I need Tmux to know when I am pasting (so it can ignore those bindings).
I thought Tmux would enable the terminal's bracketed paste mode and find
out that way, but it doesn't seem to. So setting assume-paste-time back
to its default seems to be the best solution for now.